+++ modules/actionbars/bars/bar_xp_rep.lua
@@ -296,6 +296,7 @@ BarWidget.OnEnable = function(self)
 				self.InVehicle = nil
 			end
 		end
+		self:Update()
 	end)
 	
 	self:RegisterEvent("PLAYER_ALIVE", "Update")